Credentials propagation improvements + dev docs
